fix: show coupon deductions in parabol
