+++ packages/server/graphql/public/types/User.ts
@@ -1,7 +1,14 @@
 import {isNotNull} from 'parabol-client/utils/predicates'
-import {isSuperUser} from '../../../utils/authorization'
+import {Threshold} from '../../../../client/types/constEnums'
+import fetchAllLines from '../../../billing/helpers/fetchAllLines'
+import generateInvoice from '../../../billing/helpers/generateInvoice'
+import generateUpcomingInvoice from '../../../billing/helpers/generateUpcomingInvoice'
+import getRethink from '../../../database/rethinkDriver'
+import {getUserId, isSuperUser, isUserBillingLeader} from '../../../utils/authorization'
 import getDomainFromEmail from '../../../utils/getDomainFromEmail'
 import isCompanyDomain from '../../../utils/isCompanyDomain'
+import standardError from '../../../utils/standardError'
+import {getStripeManager} from '../../../utils/stripe'
 import {UserResolvers} from '../resolverTypes'
 
 const User: UserResolvers = {
@@ -24,6 +31,32 @@ const User: UserResolvers = {
   },
   featureFlags: ({featureFlags}) => {
     return Object.fromEntries(featureFlags.map((flag) => [flag as any, true]))
+  },
+  invoiceDetails: async (_source, {invoiceId}, {authToken, dataLoader}) => {
+    const r = await getRethink()
+    const now = new Date()
+
+    const viewerId = getUserId(authToken)
+    const isUpcoming = invoiceId.startsWith('upcoming_')
+    const currentInvoice = await r.table('Invoice').get(invoiceId).default(null).run()
+    const orgId = (currentInvoice && currentInvoice.orgId) || invoiceId.substring(9) // remove 'upcoming_'
+    if (!(await isUserBillingLeader(viewerId, orgId, dataLoader))) {
+      standardError(new Error('Not organization lead'), {userId: viewerId})
+      return null
+    }
+    if (currentInvoice) {
+      const invalidAt = new Date(
+        currentInvoice.createdAt.getTime() + Threshold.UPCOMING_INVOICE_TIME_VALID
+      )
+      if (invalidAt > now) return currentInvoice
+    }
+    if (isUpcoming) {
+      return generateUpcomingInvoice(orgId, dataLoader)
+    }
+    const manager = getStripeManager()
+    const stripeLineItems = await fetchAllLines(invoiceId)
+    const invoice = await manager.retrieveInvoice(invoiceId)
+    return generateInvoice(invoice, stripeLineItems, orgId, invoiceId, dataLoader)
   }
 }
